// src/app/interfaces/reporte.ts
export interface Reporte {
    idCampus: number;
    nombreCampus: string;
    precio: number;
    descuentoJugador: number;
    formaPagoUno: string;
    formaPagoDos: string;
    formaPagoTres: string;
    nombreParticipante: string;
    primerApellido: string;
    correoParticipante: string;
    esJugador: string
}